refactor(quit): extract QuitHoldOverlay + DRY timer teardown (#306)
Simplify QuitHoldController without changing behavior.

- Extract the overlay/panel/linger/fade code into a dedicated
  `@MainActor QuitHoldOverlay` (show/update/cancelWithLinger/hide + the panel
  and SwiftUI view types). The controller is now gesture orchestration only
  (252 lines, down from ~400). Mechanical move; the `=== panel` identity guard
  in beginFadeOut and the owner-window-close observer lifecycle are preserved
  verbatim (the observer now fires an `onOwnerClose` closure back to the
  controller's `handleExternalCancellation()`).
- DRY the work-item teardown: add `cancelHoldTimer()`; `hide()`/`overlay.hide()`
  is the sole owner of `lingerWorkItem`. Drops the redundant cancels in
  `.dismiss` and `fireTermination()`.

`fireTermination()` still calls `NSApp.terminate(nil)` synchronously from the
monitor's release handler (during sendEvent) — unchanged. Decision, filter,
AppDelegate, WindowStatesManager, and all tests untouched. 23 hold-to-quit
tests green; SwiftLint + SwiftFormat clean.
